AI_SMART_KIOSK/requirements.txt
flask==3.0.3
flask-cors==4.0.1
boto3==1.34.148
botocore==1.34.148
python-dotenv==1.0.1

